update to use SAM, and add deconfigure script
