+++ .gitignore
@@ -1,5 +1,3 @@
 node_modules
 npm-debug.log
 .DS_Store
-example/lambda-function.zip
-example/lambda-invoke-response.json
